Corretta patch relativa al [bug:#285] che cancellava si tutti temporanei i file utilizzati da tcpdf nella generazione dei documenti, ma purtroppo anche quelli utilizzati in eventuali personalizzazioni dei templates. Adesso cancella solo quelli con estensione .tmp
git-svn-id: https://svn.code.sf.net/p/gazie/code/trunk@2723 0ed6177e-08db-47c9-9b37-4a62b3198fc7

+++ library/tcpdf/tcpdf.php
@@ -9994,8 +9994,12 @@ class TCPDF {
 		$this->_out($o);
 		$this->_out('%%EOF');
 		$this->state = 3; // end-of-doc
+		// GAzie patch: elimino i file che ho usato per generare il pdf ma solo se temporanei altrimenti rischio di cancellare dal filesystem del server anche quelli usati per personalizzare i templates
 		foreach ($this->imagekeys as $file) {
-			unlink($file);
+			$ex=explode('.', $file);
+			if (end($ex)=='tmp'){
+				unlink($file);
+			}
 		}
 	}
